+++ packages/relayer/src/rpcRelayer/rpcService.ts
@@ -57,12 +57,13 @@ export interface TxnReciptLog {
   data: string;
 }
 
-export interface TxnReciptResult {
+export interface TxnReceiptResult {
   txHash: string;
   index: number;
-  status: string;
+  status: number;
   revertReason?: string;
-  logs: TxnReciptLog[];
+  logs?: TxnReciptLog[];
+  receipts: TxnReceiptResult[];
 }
 
 export interface EstimateGasResult {
@@ -78,7 +79,7 @@ export interface RpcService {
   feeTokens(headers?: object): Promise<FeeTokensReturn>;
   feeOptions(feeOptionArgs: FeeOptionArgs, headers?: object): Promise<FeeOptionsReturn>;
   sendTransaction(args: PendingExecuteCallArgs, headers?: object): Promise<string>;
-  txRecipt(txHash: string, headers?: object): Promise<TxnReciptResult>;
+  txRecipt(txHash: string, headers?: object): Promise<TxnReceiptResult>;
   estimateGas(pendingExecuteCallArgs: PendingExecuteCallArgs, headers?: object): Promise<EstimateGasResult>;
   nonce(walletAddr: string, headers?: object): Promise<BigNumber>;
   metaNonce(walletAddr: string, headers?: object): Promise<BigNumber>;
@@ -99,7 +100,7 @@ export class RpcService implements RpcService {
   }
 
   async feeTokens(headers?: object): Promise<FeeTokensReturn> {
-    const res = await this.fetch(this.url("fee_tokens"), createGetPostHTTPRequest(headers));
+    const res = await this.fetch(this.url("/fee_tokens"), createGetPostHTTPRequest(headers));
     const _data = await buildResponse(res);
     return {
       isFeeRequired: <boolean>_data.isFeeRequired,
@@ -119,31 +120,31 @@ export class RpcService implements RpcService {
   async sendTransaction(args: PendingExecuteCallArgs, headers?: object): Promise<string> {
     const res = await this.fetch(this.url("/send_transaction"), createPostHTTPRequest(args, headers));
     const _data = await buildResponse(res);
-    return <string>_data.data;
+    return <string>_data;
   }
 
-  async txRecipt(txHash: string, headers?: object): Promise<TxnReciptResult> {
+  async txRecipt(txHash: string, headers?: object): Promise<TxnReceiptResult> {
     const res = await this.fetch(this.url(`/tx_receipt/${txHash}`), createGetPostHTTPRequest(headers));
     const _data = await buildResponse(res);
-    return <TxnReciptResult>_data.data;
+    return <TxnReceiptResult>_data;
   }
 
   async estimateGas(pendingExecuteCallArgs: PendingExecuteCallArgs, headers?: object): Promise<EstimateGasResult> {
     const res = await this.fetch(this.url(`/estimate_gas`), createPostHTTPRequest(pendingExecuteCallArgs, headers));
     const _data = await buildResponse(res);
-    return _data.data;
+    return _data;
   }
 
   async nonce(walletAddr: string, headers?: object): Promise<BigNumber> {
     const res = await this.fetch(this.url(`/nonce/${walletAddr}`), createGetPostHTTPRequest(headers));
     const _data = await buildResponse(res);
-    return BigNumber.from(_data.data);
+    return BigNumber.from(_data);
   }
 
   async metaNonce(walletAddr: string, headers?: object): Promise<BigNumber> {
     const res = await this.fetch(this.url(`/meta_nonce/${walletAddr}`), createGetPostHTTPRequest(headers));
     const _data = await buildResponse(res);
-    return BigNumber.from(_data.data);
+    return BigNumber.from(_data);
   }
 }
 
@@ -179,11 +180,11 @@ const buildResponse = (res: Response): Promise<any> =>
       } as WebRPCError;
     }
 
-    if (!res.ok) {
+    if (!res.ok || data.statusCode !== 200) {
       throw data; // webrpc error response
     }
 
-    return data;
+    return data.data;
   });
 
 export type Fetch = (input: RequestInfo, init?: RequestInit) => Promise<Response>;