+++ packages/deployer/CHANGELOG.md
@@ -1,5 +1,46 @@
 # @unipasswallet/deployer
 
+## 0.0.1-alpha.11
+
+### Patch Changes
+
+- @unipasswallet/keys: update json parse
+  @unipasswallet/relayer: update receipt
+- Updated dependencies
+  - @unipasswallet/utils@0.0.1-alpha.15
+
+## 0.0.1-alpha.9
+
+### Patch Changes
+
+- update versions
+- Updated dependencies
+  - @unipasswallet/utils@0.0.1-alpha.13
+
+## 0.0.1-alpha.8
+
+### Patch Changes
+
+- update versions
+- Updated dependencies
+  - @unipasswallet/utils@0.0.1-alpha.12
+
+## 0.0.1-alpha.7
+
+### Patch Changes
+
+- update versions
+- Updated dependencies
+  - @unipasswallet/utils@0.0.1-alpha.11
+
+## 0.0.1-alpha.6
+
+### Patch Changes
+
+- update versions
+- Updated dependencies
+  - @unipasswallet/utils@0.0.1-alpha.10
+
 ## 0.0.1-alpha.5
 
 ### Patch Changes