+++ packages/dkim-base/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @unipasswallet/dkim-base
 
+## 0.0.1-alpha.15
+
+### Patch Changes
+
+- @unipasswallet/keys: update json parse
+  @unipasswallet/relayer: update receipt
+
+## 0.0.1-alpha.13
+
+### Patch Changes
+
+- update versions
+
+## 0.0.1-alpha.12
+
+### Patch Changes
+
+- update versions
+
+## 0.0.1-alpha.11
+
+### Patch Changes
+
+- update versions
+
+## 0.0.1-alpha.10
+
+### Patch Changes
+
+- update versions
+
 ## 0.0.1-alpha.9
 
 ### Patch Changes