+++ packages/dkim/CHANGELOG.md
@@ -1,5 +1,46 @@
 # @unipasswallet/dkim
 
+## 0.0.1-alpha.14
+
+### Patch Changes
+
+- @unipasswallet/keys: update json parse
+  @unipasswallet/relayer: update receipt
+- Updated dependencies
+  - @unipasswallet/dkim-base@0.0.1-alpha.15
+
+## 0.0.1-alpha.13
+
+### Patch Changes
+
+- update versions
+- Updated dependencies
+  - @unipasswallet/dkim-base@0.0.1-alpha.13
+
+## 0.0.1-alpha.12
+
+### Patch Changes
+
+- update versions
+- Updated dependencies
+  - @unipasswallet/dkim-base@0.0.1-alpha.12
+
+## 0.0.1-alpha.11
+
+### Patch Changes
+
+- update versions
+- Updated dependencies
+  - @unipasswallet/dkim-base@0.0.1-alpha.11
+
+## 0.0.1-alpha.10
+
+### Patch Changes
+
+- update versions
+- Updated dependencies
+  - @unipasswallet/dkim-base@0.0.1-alpha.10
+
 ## 0.0.1-alpha.9
 
 ### Patch Changes