+++ packages/wallet/CHANGELOG.md
@@ -1,5 +1,76 @@
 # @unipasswallet/wallet
 
+## 0.0.1-alpha.15
+
+### Patch Changes
+
+- @unipasswallet/keys: update json parse
+  @unipasswallet/relayer: update receipt
+- Updated dependencies
+  - @unipasswallet/relayer@0.0.1-alpha.10
+  - @unipasswallet/abi@0.0.1-alpha.15
+  - @unipasswallet/dkim-base@0.0.1-alpha.15
+  - @unipasswallet/keys@0.0.1-alpha.15
+  - @unipasswallet/network@0.0.1-alpha.15
+  - @unipasswallet/transactions@0.0.1-alpha.14
+  - @unipasswallet/utils@0.0.1-alpha.15
+
+## 0.0.1-alpha.14
+
+### Patch Changes
+
+- update versions
+- Updated dependencies
+  - @unipasswallet/abi@0.0.1-alpha.13
+  - @unipasswallet/dkim-base@0.0.1-alpha.13
+  - @unipasswallet/keys@0.0.1-alpha.13
+  - @unipasswallet/network@0.0.1-alpha.13
+  - @unipasswallet/relayer@0.0.1-alpha.9
+  - @unipasswallet/transactions@0.0.1-alpha.13
+  - @unipasswallet/utils@0.0.1-alpha.13
+
+## 0.0.1-alpha.13
+
+### Patch Changes
+
+- update versions
+- Updated dependencies
+  - @unipasswallet/abi@0.0.1-alpha.12
+  - @unipasswallet/dkim-base@0.0.1-alpha.12
+  - @unipasswallet/keys@0.0.1-alpha.12
+  - @unipasswallet/network@0.0.1-alpha.12
+  - @unipasswallet/relayer@0.0.1-alpha.8
+  - @unipasswallet/transactions@0.0.1-alpha.12
+  - @unipasswallet/utils@0.0.1-alpha.12
+
+## 0.0.1-alpha.12
+
+### Patch Changes
+
+- update versions
+- Updated dependencies
+  - @unipasswallet/abi@0.0.1-alpha.11
+  - @unipasswallet/dkim-base@0.0.1-alpha.11
+  - @unipasswallet/keys@0.0.1-alpha.11
+  - @unipasswallet/network@0.0.1-alpha.11
+  - @unipasswallet/relayer@0.0.1-alpha.7
+  - @unipasswallet/transactions@0.0.1-alpha.11
+  - @unipasswallet/utils@0.0.1-alpha.11
+
+## 0.0.1-alpha.11
+
+### Patch Changes
+
+- update versions
+- Updated dependencies
+  - @unipasswallet/abi@0.0.1-alpha.10
+  - @unipasswallet/dkim-base@0.0.1-alpha.10
+  - @unipasswallet/keys@0.0.1-alpha.10
+  - @unipasswallet/network@0.0.1-alpha.10
+  - @unipasswallet/relayer@0.0.1-alpha.6
+  - @unipasswallet/transactions@0.0.1-alpha.10
+  - @unipasswallet/utils@0.0.1-alpha.10
+
 ## 0.0.1-alpha.10
 
 ### Patch Changes